# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditLogOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_sensitive": {
            "title": "Is Sensitive",
            "type": "boolean"
          },
          "record_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          }
        },
        "required": [
          "id",
          "actor_id",
          "action",
          "table_name",
          "record_id",
          "details",
          "is_sensitive",
          "created_at"
        ],
        "title": "AuditLogOut",
        "type": "object"
      },
...
      "ContactMessageIn": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "contact_name": {
            "minLength": 1,
            "title": "Contact Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "type": {
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "contact_name",
          "email",
          "message"
        ],
        "title": "ContactMessageIn",
        "type": "object"
      },
      "ContactMessageOut": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "contact_name": {
            "title": "Contact Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "company",
          "contact_name",
          "email",
          "phone",
          "message",
          "status",
          "created_at"
        ],
        "title": "ContactMessageOut",
        "type": "object"
      },
      "ContactStatusUpdate": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ContactStatusUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationFeed": {
        "description": "O que o sino precisa em uma chamada só: a lista e o número.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread_count"
        ],
        "title": "NotificationFeed",
        "type": "object"
      },
...
      "PlatformUpdateIn": {
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "PlatformUpdateIn",
        "type": "object"
      },
      "PlatformUpdateOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft": {
            "title": "Draft",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notified_count": {
            "title": "Notified Count",
            "type": "integer"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "draft",
          "notified_count",
          "published_at",
          "created_at"
        ],
        "title": "PlatformUpdateOut",
        "type": "object"
      },
...
      "ResumoDeAuditoriaOut": {
        "description": "Os cartões e o gráfico do painel de auditoria (#34 a #37 da conferência).",
        "properties": {
          "atividade": {
            "items": {
              "$ref": "#/components/schemas/DiaDeAuditoriaOut"
            },
            "title": "Atividade",
            "type": "array"
          },
          "hoje": {
            "title": "Hoje",
            "type": "integer"
          },
          "mais_ativo_acoes": {
            "title": "Mais Ativo Acoes",
            "type": "integer"
          },
          "mais_ativo_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mais Ativo Nome"
          },
          "sensiveis_sete_dias": {
            "title": "Sensiveis Sete Dias",
            "type": "integer"
          },
          "sete_dias": {
            "title": "Sete Dias",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "hoje",
          "sete_dias",
          "sensiveis_sete_dias",
          "mais_ativo_nome",
          "mais_ativo_acoes",
          "atividade"
        ],
        "title": "ResumoDeAuditoriaOut",
        "type": "object"
      },
...
      "SettingOut": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "persisted": {
            "title": "Persisted",
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "value",
          "updated_at",
          "updated_by",
          "persisted"
        ],
        "title": "SettingOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API multi-tenant de feedback 360 e avaliação de clientes.",
    "title": "A&S Feedback Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit-logs": {
      "get": {
        "operationId": "list_audit_logs_api_v1_audit_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditLogOut"
                  },
                  "title": "Response List Audit Logs Api V1 Audit Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit Logs",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/audit-logs/summary": {
      "get": {
        "description": "Cartões e gráfico do painel de auditoria.\n\nRota separada da listagem porque responde outra pergunta: a listagem pagina o\ndetalhe, esta agrega o todo. Juntar as duas obrigaria a recalcular os agregados a\ncada página virada.",
        "operationId": "audit_summary_api_v1_audit_logs_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoDeAuditoriaOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Summary",
        "tags": [
          "engagement"
        ]
      }
    },
...
    "/api/v1/contact-messages": {
      "get": {
        "operationId": "list_contact_messages_api_v1_contact_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContactMessageOut"
                  },
                  "title": "Response List Contact Messages Api V1 Contact Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Contact Messages",
        "tags": [
          "engagement"
        ]
      },
      "post": {
        "operationId": "create_contact_message_api_v1_contact_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactMessageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Contact Message",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/contact-messages/{message_id}": {
      "patch": {
        "operationId": "change_contact_status_api_v1_contact_messages__message_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Contact Status",
        "tags": [
          "engagement"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "description": "Lista e contagem em uma chamada só — é o que o sino consome a cada página.",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationFeed"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "operationId": "mark_all_notifications_read_api_v1_notifications_read_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark All Notifications Read",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_notification_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Notification Read",
        "tags": [
          "engagement"
        ]
      }
    },
...
    "/api/v1/platform-updates": {
      "get": {
        "description": "Todo mundo vê os publicados; admin/RH também vê os rascunhos.",
        "operationId": "list_platform_updates_api_v1_platform_updates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlatformUpdateOut"
                  },
                  "title": "Response List Platform Updates Api V1 Platform Updates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Platform Updates",
        "tags": [
          "engagement"
        ]
      },
      "post": {
        "operationId": "create_platform_update_api_v1_platform_updates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUpdateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Platform Update",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/platform-updates/{update_id}/publish": {
      "post": {
        "description": "Publicar enfileira uma mensagem por destinatário, na mesma transação (AD-04).\n\nSe o worker estiver fora do ar, a publicação acontece do mesmo jeito e as mensagens\nesperam na fila — que é exatamente o cenário \"efeito auxiliar não desfaz operação\nprincipal\" de PAR-07.",
        "operationId": "publish_platform_update_api_v1_platform_updates__update_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "update_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Update Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUpdateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Platform Update",
        "tags": [
          "engagement"
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "description": "Leitura é de qualquer autenticado: os toggles decidem o que a navegação mostra.\n\nEscrita continua restrita a admin/RH — e a autorização final de cada recurso é\nchecada no servidor, nunca por escondido de menu (BR-MIGRAR-016).",
        "operationId": "list_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SettingOut"
                  },
                  "title": "Response List Settings Api V1 Settings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Settings",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/v1/settings/{key}": {
      "put": {
        "operationId": "update_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Setting",
        "tags": [
          "engagement"
        ]
      }
    },
...
  }
}